Arreglos de conflictos con otra rama
